refactor: input 부분을 자식 컴포넌트로 분리
