+++ src/features/goal/components/detail/Tasks.tsx
@@ -1,3 +1,6 @@
+import type { Dispatch } from 'react';
+import type { SetStateAction } from 'jotai';
+
 import PlusIcon from '@/assets/icons/plus.svg';
 import { Typography } from '@/components';
 import { GoalDetail } from '@/components/molecules';
@@ -5,9 +8,10 @@ import type { GoalTasksProps } from '@/hooks/reactQuery/goal/useGetGoal';
 
 interface TasksProps {
   tasks: GoalTasksProps[];
-  onOpenInput: (status: boolean) => void;
+  onOpenInput: Dispatch<SetStateAction<boolean>>;
 }
 
+// TODO : 목표 달성, 더보기 버튼 클릭 이벤트 등록
 export const Tasks = ({ tasks, onOpenInput }: TasksProps) => {
   return (
     <div className="flex flex-col gap-4xs">
